// src/Data-Test/Note-Creation-Interface.ts
export interface NoteDataInterface {
    title: string;
    secondaryTitle?: string;
    subtitle?: string;
    halfTitle?: string;
    body: string;
    tags?: string[];
    hiddenTags?: string[];
    summary?: string;
    authorName?: string;
    authorDescription?: string;
    listicleTitle?: string[];
    listicleBody?: string[];
}